# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/workouts": {
      "get": {
        "operationId": "WorkoutsController_getCompletedWorkouts_v1",
        "parameters": [
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workouts"
        ]
      },
      "post": {
        "operationId": "WorkoutsController_createDraftWorkout_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workouts"
        ]
      }
    },
    "/v1/workouts/count": {
      "get": {
        "operationId": "WorkoutsController_getCompletedWorkoutsCount_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workouts"
        ]
      }
    },
    "/v1/workouts/calendar": {
      "get": {
        "operationId": "WorkoutsController_getWorkoutCalendar_v1",
        "parameters": [
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workouts"
        ]
      }
    },
    "/v1/workouts/stats": {
      "get": {
        "operationId": "WorkoutsController_getWorkoutStats_v1",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workouts"
        ]
      }
    },
    "/v1/workouts/active": {
      "get": {
        "operationId": "WorkoutsController_getActiveWorkout_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workouts"
        ]
      },
      "post": {
        "operationId": "WorkoutsController_createActiveWorkout_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workouts"
        ]
      },
      "delete": {
        "operationId": "WorkoutsController_deleteActiveWorkout_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workouts"
        ]
      }
    },
    "/v1/workouts/{workoutId}": {
      "get": {
        "operationId": "WorkoutsController_getWorkout_v1",
        "parameters": [
          {
            "name": "workoutId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workouts"
        ]
      },
      "patch": {
        "operationId": "WorkoutsController_updateWorkout_v1",
        "parameters": [
          {
            "name": "workoutId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workouts"
        ]
      },
      "delete": {
        "operationId": "WorkoutsController_deleteWorkout_v1",
        "parameters": [
          {
            "name": "workoutId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workouts"
        ]
      }
    },
    "/v1/workouts/active/pause": {
      "patch": {
        "operationId": "WorkoutsController_pauseActiveWorkout_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workouts"
        ]
      }
    },
    "/v1/workouts/active/resume": {
      "patch": {
        "operationId": "WorkoutsController_resumeActiveWorkout_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workouts"
        ]
      }
    },
    "/v1/workouts/{workoutId}/complete": {
      "post": {
        "operationId": "WorkoutsController_completeWorkout_v1",
        "parameters": [
          {
            "name": "workoutId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workouts"
        ]
      }
    },
    "/v1/workouts/{workoutId}/workoutExercises": {
      "post": {
        "operationId": "WorkoutsController_createWorkoutExercise_v1",
        "parameters": [
          {
            "name": "workoutId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkoutExerciseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workouts"
        ]
      }
    },
    "/v1/workouts/{workoutId}/workoutExercises/{workoutExerciseId}": {
      "patch": {
        "operationId": "WorkoutsController_updateWorkoutExercise_v1",
        "parameters": [
          {
            "name": "workoutId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "workoutExerciseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkoutExerciseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workouts"
        ]
      },
      "delete": {
        "operationId": "WorkoutsController_deleteWorkoutExercise_v1",
        "parameters": [
          {
            "name": "workoutId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "workoutExerciseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workouts"
        ]
      }
    },
    "/v1/workouts/{workoutId}/workoutExercises/{workoutExerciseId}/sets": {
      "get": {
        "operationId": "WorkoutsController_getWorkoutExerciseSets_v1",
        "parameters": [
          {
            "name": "workoutId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "workoutExerciseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workouts"
        ]
      },
      "post": {
        "operationId": "WorkoutsController_createWorkoutSet_v1",
        "parameters": [
          {
            "name": "workoutId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "workoutExerciseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkoutSetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workouts"
        ]
      }
    },
    "/v1/workouts/{workoutId}/workoutExercises/{workoutExerciseId}/sets/{setId}": {
      "patch": {
        "operationId": "WorkoutsController_updateWorkoutSet_v1",
        "parameters": [
          {
            "name": "workoutId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "workoutExerciseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "setId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkoutSetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workouts"
        ]
      },
      "delete": {
        "operationId": "WorkoutsController_deleteWorkoutSet_v1",
        "parameters": [
          {
            "name": "workoutId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "workoutExerciseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "setId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workouts"
        ]
      }
    },
...
    "/v1/exercises/{exerciseId}/workouts": {
      "get": {
        "operationId": "ExercisesController_getExerciseWorkouts_v1",
        "parameters": [
          {
            "name": "exerciseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Exercises"
        ]
      }
    },
...
  },
  "info": {
    "title": "NextLift API",
    "description": "API documentation for the NextLift application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateWorkoutDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateWorkoutDto": {
        "type": "object",
        "properties": {}
      },
      "CreateWorkoutExerciseDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateWorkoutExerciseDto": {
        "type": "object",
        "properties": {}
      },
      "CreateWorkoutSetDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateWorkoutSetDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}